% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.snowReservoir.R
\name{init.snowReservoir}
\alias{init.snowReservoir}
\title{Initialize information related to snow}
\usage{
init.snowReservoir(method = NULL, path = NULL, snomag = NULL,
  swe_h = NULL, middelsca = NULL, snofritt = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}

\item{snomag}{snow mag, NEED MORE EXPLANATION}

\item{swe_h}{swe_h, NEED MORE EXPLANATION}

\item{middelsca}{averaged sca}

\item{snofritt}{snowfree, NEED MORE EXPLANATION}
}
\description{
Initialize the information related to snow
}
\examples{
\dontrun{
init.snowReservoir()
}
}
\keyword{snow}

