% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoir.R
\name{reservoir}
\alias{reservoir}
\title{Compute the snow reservoir}
\usage{
reservoir(snow)
}
\arguments{
\item{snow}{snow}
}
\description{
Compute the snow reservoir
}
\examples{
\dontrun{
reservoir()
}
}
\keyword{snow}

