% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snogamma_varc3_SCAupdate.R
\name{snogamma}
\alias{snogamma}
\title{snogamma}
\usage{
snogamma(PRX, PSX, MWX, scax, scaobx, spdx, wcdx, prox, nsnox, alfax, nyx,
  alfa0x, ny0x, ac, hc, UP)
}
\arguments{
\item{PRX}{nedbor som regn}

\item{PSX}{nedbor som sno}

\item{MWX}{potenial melt/refreeze(negative) mm}

\item{scax}{dekningsgrad (i hver hoydesone)(array(idim))}

\item{scaobx}{snow coverage observation}

\item{spdx}{Snow Water Equivalent in BV. NEED MORE EXPLANATION}

\item{wcdx}{Free Water in BV. NEED MORE EXPLANATION}

\item{prox}{Max Liquid water content in snow. Percentage of Snow Water Equivalent}

\item{nsnox}{number of event. NEED MORE EXPLANATION}

\item{alfax}{alfa parameter i nedbor gamma fordeling}

\item{nyx}{ny parameter i nedbor gamma fordeling}

\item{alfa0x}{alfa parameter i nedbor gamma fordeling}

\item{ny0x}{ny parameter i nedbor gamma fordeling}

\item{ac}{NEED MORE EXPLANATION}

\item{hc}{NEED MORE EXPLANATION}

\item{UP}{look at scaobx or not}
}
\description{
snogamma
NEED MORE EXPLANATION
WARNING:
Some input parameters described in the source file
are however not involved:
- cc   :temporal correlasjon mellom hendelser
- pro  :max prosent fritt vann i sno
- idim :no hoydesone'
HUSK AT SPD, WCD ER UBETINGEDE VERDIER. MA GANGES MED PPA FOR A FA AREALVERDIER!!!
og Output:
ISOIL            :runoff til grunn eller overflate avrenningn UBV
SPD  	      :swe BV
WCD              :fritt vann i snopakke BV
nsno             :antall hendelser
sca              :oppdatert snodekning
hc og ac bekriver sammenhengen mellom  areelt middel og standdardavvik av nedbor se BREMS
}
\examples{
\dontrun{
snogamma()
}
}
\keyword{snow}

