% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateLZ.R
\name{stateLZ}
\alias{stateLZ}
\title{Update the information related to snow at one level zone}
\usage{
stateLZ(htemp, hprecip, scaob, modelSnow, modelPrecipLZ, gca, sca, spd, wcd,
  nsno, alfa, ny)
}
\arguments{
\item{htemp}{temperature for each elevation zone}

\item{hprecip}{precipitation for each elevation zone}

\item{scaob}{snow coverage observation}

\item{modelSnow}{list of parameters related to snow
list(nbLevelZone,unitsnow,n0,Ws,TS,CX,CFR,CGLAC,gca,UP)}

\item{modelPrecipLZ}{list of parameters related to precipLZ
list(nbLevelZone,Plr,hfelt,midmetp)}

\item{gca}{fraction of glacier pr elevation zone, fraction of glacier covered area pr elevation zone}

\item{sca}{snow coverage (NEED MORE EXPLANATION)}

\item{spd}{Snow Water Equivalent in BV (NEED MORE EXPLANATION)}

\item{wcd}{Free Water in BV (NEED MORE EXPLANATION)}

\item{nsno}{number of event (NEED MORE EXPLANATION)}

\item{alfa}{alfa parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)}

\item{ny}{ny parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)}
}
\description{
Update the information related to snow at one level zone
}
\examples{
\dontrun{
stateLZ()
}
}
\keyword{snow}

