% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.snowReservoir.R
\name{stateX.snowReservoir}
\alias{stateX.snowReservoir}
\title{Compute the snow reservoir}
\usage{
stateX.snowReservoir(snow)
}
\arguments{
\item{snow}{snow}
}
\description{
Compute the snow reservoir
}
\examples{
\dontrun{
stateX.snowReservoir()
}
}
\keyword{snow}

