% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varc4.R
\name{varc}
\alias{varc}
\title{Update dynamical values of nu and alfa}
\usage{
varc(ppa, ac, drange, nudyn, alphadyn, sdnux, sdalphax, nnn, u, n, redsca)
}
\arguments{
\item{ppa}{SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units}

\item{ac}{(previously alfanull) parameters of relationship mean standarddeviation of conditional precip.}

\item{drange}{parameters of relationship mean standarddeviation of conditional precip.}

\item{nudyn}{initial distr parameters, current distr parameters}

\item{alphadyn}{initial distr parameters, current distr parameters}

\item{sdnux}{(ny0) initial distr parameters, current distr parameters}

\item{sdalphax}{(alfa0) initial distr parameters, current distr parameters}

\item{nnn}{SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units}

\item{u}{SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units}

\item{n}{SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units}

\item{redsca}{SCA,relative reduction of SCA, previously accumulated units,current accu units and melt units}
}
\value{
a list of three updated values: nudyn,alphadyn,nnn
}
\description{
The function \code{varc()} updates and returns dynamical values of
nu and alfa, of the snowroutine. The correlation coefficient is evaluated
as the sum of the correlation N*N, 2*N*u and u*u matrices
Should be a decreasing value with n, large n small c
All variables must be conditional (i.e. calculated on snowcovered area)
Note the distinction between conditional and unconditional units of melt
18.03.2015 Recodet using only the correlation function. Alfanull ( sdalphax) and nynull( sdnux) are input
}
\examples{
\dontrun{
varc()
}
}
\keyword{snow}

