% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.snow.R, R/init.snowReservoir.R
\name{init.snow}
\alias{init.snow}
\title{Initialize information related to snow}
\usage{
init.snow(method = NULL, path = NULL, snomag = NULL, swe_h = NULL,
  middelsca = NULL, snofritt = NULL)

init.snow(method = NULL, path = NULL, snomag = NULL, swe_h = NULL,
  middelsca = NULL, snofritt = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}

\item{snomag}{snow mag, NEED MORE EXPLANATION}

\item{swe_h}{swe_h, NEED MORE EXPLANATION}

\item{middelsca}{averaged sca}

\item{snofritt}{snowfree, NEED MORE EXPLANATION}

\item{isoil}{precipitation and/or snowmelt from the elevation zones, vector(level zone)}

\item{gisoil}{glaciermelt from the elevation zones, vector(level zone)}

\item{sca}{snow coverage (NEED MORE EXPLANATION)}

\item{spd}{Snow Water Equivalent in BV (NEED MORE EXPLANATION)}

\item{wcd}{Free Water in BV (NEED MORE EXPLANATION)}

\item{nsno}{number of event (NEED MORE EXPLANATION)}

\item{alfa}{alfa parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)}

\item{ny}{ny parameter i nedbor gamma fordeling (NEED MORE EXPLANATION)}

\item{snowfree}{percentage of snowfree area}

\item{method}{method for the initialization, "load", "source", "manual"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}
}
\description{
Initialize the information related to snow

Initialize the information related to snow
}
\examples{
\dontrun{
init.snow()
}
\dontrun{
init.snow()
}
}
\keyword{snow}

